Using Redux-Thunk and actions spread across multiple files.
