+++ Redux/src/containers/Counter/Counter.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { increment, decrement, add, subtract, storeResult, deleteResult } from '../../store/actions';
+import { increment, decrement, add, subtract, storeResult, deleteResult } from '../../store/actions/index';
 
 import CounterControl from '../../components/CounterControl/CounterControl';
 import CounterOutput from '../../components/CounterOutput/CounterOutput';
