+++ Redux/src/store/actions/actionTypes.js
@@ -1,0 +1,15 @@
+export const INCREMENT = 'INCREMENT';
+export const DECREMENT = 'DECREMENT';
+export const ADD = 'ADD';
+export const SUBTRACT = 'SUBTRACT';
+export const STORE_RESULT = 'STORE_RESULT';
+export const DELETE_RESULT = 'DELETE_RESULT';
+
+
+/*
+
+    For larger applications, these prevent spelling mistakes of the strings. They must match.
+    Prevents RUNTIME errors. Instead, there will be a compile error and can be fixed easy.
+
+*/
+
