+++ Redux/src/store/actions/index.js
@@ -1,0 +1,11 @@
+export {
+    add,
+    subtract,
+    increment,
+    decrement
+} from './counter';
+
+export {
+    storeResult,
+    deleteResult
+} from './result';
